workers/requirements.txt
psycopg2-binary==2.9.10
requests==2.32.3
Pillow==11.1.0
numpy<2.1
flask==3.1.0
tensorflow==2.19.0
anthropic>=0.40.0
pytest>=8.0
